+++ instrumentation/micrometer/micrometer-1.5/testing/src/main/java/io/opentelemetry/instrumentation/micrometer/v1_5/AbstractDistributionSummaryTest.java
@@ -61,21 +61,6 @@ public abstract class AbstractDistributionSummaryTest {
                                         .hasAttributesSatisfyingExactly(
                                             equalTo(stringKey("tag"), "value"))
                                         .hasBucketBoundaries(NO_BUCKETS))));
-    testing()
-        .waitAndAssertMetrics(
-            INSTRUMENTATION_NAME,
-            metric ->
-                metric
-                    .hasName("testSummary.max")
-                    .hasDescription("This is a test distribution summary")
-                    .hasDoubleGaugeSatisfying(
-                        gauge ->
-                            gauge.hasPointsSatisfying(
-                                point ->
-                                    point
-                                        .hasValue(4)
-                                        .hasAttributesSatisfyingExactly(
-                                            equalTo(stringKey("tag"), "value")))));
 
     // micrometer gauge histogram is not emitted
     testing()
@@ -86,8 +71,7 @@ public abstract class AbstractDistributionSummaryTest {
     Metrics.globalRegistry.remove(summary);
 
     // then
-    // Histogram is synchronous and returns previous value after removal, max is asynchronous and is
-    // removed completely.
+    // Histogram is synchronous and returns previous value after removal.
     testing()
         .waitAndAssertMetrics(
             INSTRUMENTATION_NAME,